refactor: react-doctor 100点対応
