+++ .reactdoctor.json
@@ -1,0 +1,8 @@
+{
+  "rules": {
+    "exhaustive-deps": "off",
+    "async-await-in-loop": "off",
+    "prefer-html-dialog": "off",
+    "no-multi-comp": "off"
+  }
+}
