+++ pnpm-workspace.yaml
@@ -1,2 +1,4 @@
 allowBuilds:
   lefthook: true
+minimumReleaseAge: 10080
+trustPolicy: no-downgrade
