+++ src/constants/storage.ts
@@ -1,0 +1,3 @@
+export const STORAGE_KEYS = {
+  DECK_LIST: "pokemonTcgDeckList",
+} as const;
